requirements.txt
fastapi
langchain_core
langchain_openai
librosa
numpy
pandas
scipy
soundfile
torch
transformers
